pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unitree-go2-robot-controller"
version = "0.1.0"
description = "Minimal RAI ReAct voice agent for Unitree Go2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "go2-webrtc-connect",
    "langchain-core",
    "langchain-openai",
    "langgraph",
    "numpy",
    "openai[realtime]",
    "Pillow",
    "pydantic",
    "python-dotenv",
    "sounddevice",
]

[project.scripts]
unitree-go2-controller = "unitree_go2_robot_controller.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
